requirements_backup.txt
alabaster==0.7.12
Babel==2.9.1
bleach==4.1.0
build==0.6.0.post1
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
cycler==0.10.0
Cython==0.29.24
docutils==0.17.1
idna==3.2
imagesize==1.2.0
importlib-metadata==4.8.1
Jinja2==3.0.1
keyring==23.2.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.1
opencv-python==4.5.3.56
packaging==21.0
pep517==0.11.0
Pillow==8.3.1
pkginfo==1.7.1
pycocotools==2.0.2
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
pywin32-ctypes==0.2.0
readme-renderer==29.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
six==1.16.0
snowballstemmer==2.1.0
sortedcontainers==2.4.0
Sphinx==4.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tomli==1.2.1
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.2
twine==3.4.2
typing-extensions==3.10.0.0
urllib3==1.26.6
webencodings==0.5.1
zipp==3.5.0
